rafactor: Split test cases in separate files
